Remove obsolete consumer api integration tests
